Temp Memory leaks fix, Change Int32 to UInt8 for asset properties, Dismiss dialog from first click, rejection tool, used addresses under expert mode, fix recovery issue for confirmed txs ...

+++ ravenwallet/src/Models/Asset/Asset.swift
@@ -12,9 +12,9 @@ struct Asset {
     var idAsset: Int
     var name: String
     var amount: Satoshis
-    var units:Int
-    var reissubale:Int
-    var hasIpfs:Int
+    var units:UInt8
+    var reissubale:UInt8
+    var hasIpfs:UInt8
     var ipfsHash:String
     var ownerShip:Int
     var hidden:Int
